Remove the need for the action
